add shareable video url and add description column
